Modal Async/Await Done

+++ assets/js/app.js
@@ -110,49 +110,24 @@ function renderScreen() {
 //inclusão de animações de movimento/alteração de cor do elemento, atualização do valor total da compra.
 var boxChecked = async function () {
   var parent = this.parentNode; //Obtém a referência do componente pai para inclusão das animações/alterações de background color
-
   var id = this.getAttribute("id"); // variável id criada para receber o id do componente 'checked'
-
   id = id.replace("chk-", ""); //Remove o prefixo determinado para obter o mesmo ID do array/localStorage
+  idItem = id;
 
-  arrItem = getList(); //Atualiza o array com base no localStorage
-
-  let result = arrItem.filter(function (el) {
-    return el.id == id; //filtra o array para obter um novo array apenas com o ID identificado
-  });
-
-  for (let element of result) {
-    let index = arrItem.indexOf(element);
-    arrItem[index].checked = this.checked; //altera o checked do valor do array com base no elemento marcado
-    !this.checked ? (arrItem[index].value = 0) : "";
-  }
-
-  localStorage.setItem(listLocalStorage, JSON.stringify(arrItem)); //altera o checked do valor do localStorage com base no elemento marcado
-
+  this.checked ? await hiddenModal(false) : addValueArr(id,0,false);
 
   //verifica itens marcados e adiciona/remove classes para animação e cor no elemento, a depender do estado.
   if (this.checked) {
-    console.log('teste1')
-
-    return new Promise(resolve => { hiddenModal(false)})
- 
-    console.log('teste2')
     parent.classList.add("checked", "animate__animated", "animate__pulse");
     spanModal.innerText = "";
-    idItem = id;
-    modal.style.display = "block";
-    valueInputModal.focus();
+    idItem = "";
   } else {
     parent.classList.remove("checked", "animate__animated", "animate__pulse");
     parent.classList.add("noChecked");
-
     //realiza a atualização do valor total em tela - é reexecutada a função, pois para os efeitos de animação ocorrerem, a tela não pode ser renderizada novamente
-    sumTotal();
     //limpa o iD para o caso de um novo item ser marcado
     idItem = "";
   }
-  //
-  //
 };
 
 //Função chamada para permitir a edição do item com o duploclick no elemento
@@ -178,8 +153,7 @@ var editItem = function () {
     arrItem[index].text = text; //atualiza o array com informações do item editado
   }
 
-  localStorage.setItem(listLocalStorage, JSON.stringify(arrItem)); //Atualiza o localstorage a partir do array editado
-  renderScreen(); // Renderiza a tela com as novas informações
+  updateLocalStorage()
 };
 
 //Função para remover item do Array/Localstorage
@@ -198,8 +172,7 @@ var removeItem = function () {
     arrItem.splice(index, 1);
   }
 
-  localStorage.setItem(listLocalStorage, JSON.stringify(arrItem)); //Atualiza a lista no localStorage após remoção do item
-  renderScreen(); // Renderiza a tela
+  updateLocalStorage()
 };
 
 // Função para gerar novo ID e adicionar item ao array
@@ -221,10 +194,14 @@ const getList = function () {
 //Renderizar itens na tela no load do formulário
 renderScreen();
 
-//Captura o click no modal e realiza a atualização do valor do item no array/LocalStorage
+
+function updateLocalStorage(){
+  localStorage.setItem(listLocalStorage, JSON.stringify(arrItem)); //
+  renderScreen(); // Render Screen
+}
 
 // Variavel funcional para adicionar valor no  array/localstorage
-var addValueArr = (idItem, valor) => {
+var addValueArr = (idItem, valor, checkedValue) => {
   let result = arrItem.filter(function (el) {
     return el.id == idItem;
   });
@@ -232,10 +209,10 @@ var addValueArr = (idItem, valor) => {
   for (let element of result) {
     let index = arrItem.indexOf(element);
     arrItem[index].value = valor;
+    arrItem[index].checked = checkedValue;
+    !checkedValue ? (arrItem[index].value = 0) : "";
   }
-
-  localStorage.setItem(listLocalStorage, JSON.stringify(arrItem)); //
-  renderScreen(); // Render Screen
+  updateLocalStorage()
 };
 
 //Captura evento submit no formulário para adição de novos itens
@@ -251,19 +228,22 @@ formSubmit.addEventListener("submit", (e) => {
   }
 
   addItem(inputItem.value.trim(), 0, 0); //função para adição de novos itens
-  localStorage.setItem(listLocalStorage, JSON.stringify(arrItem)); //atualiza localstorage com informações do array
+  updateLocalStorage()
   inputItem.value = ""; //limpa o campo após salvar dados
   inputItem.focus(); //atualiza o foco no campo input para inserção de novos itens
-  renderScreen(); //renderiza a tela com itens adicionados
 });
 
 // Função para deletar todos os itens do Array/Objeto atualizando a tela
 function deleteAllItems() {
+  const validateDelete = confirm('Você realmente deseja deletar toda a lista de compras?')
+
+  if (validateDelete){
   localStorage.clear();
   arrItem = [];
   renderScreen();
   deleteAll.hidden = true;
   deleteChecked.hidden = true;
+  }
 }
 
 //função para deletar apenas itens checkados do array/localstorage ao fim renderizando a tela;
@@ -279,8 +259,7 @@ function deleteCheckeds() {
     arrItem.splice(index, 1);
   }
 
-  localStorage.setItem(listLocalStorage, JSON.stringify(arrItem)); //
-  renderScreen(); // Render Screen
+  updateLocalStorage()
 }
 
 //Função para soma do total com base no array atualizado pelo localstorage
@@ -316,21 +295,36 @@ function applyHidden(value, render) {
   render ? ((deleteAll.hidden = value), (deleteChecked.hidden = value)) : "";
 }
 
-const hiddenModal = async (value) => {
-  errModal.hidden = false;
-  new Promise((resolve,reject) => {
+async function hiddenModal(value) {
+  return new Promise((resolve) => {
+    let stop = false;
+    errModal.hidden = !value;
+    valueInputModal.focus();
+    modal.style.display = "block";
+
     valueBtnModal.addEventListener("click", () => {
       if (valueInputModal.value) {
         valorModal = parseFloat(valueInputModal.value);
-        addValueArr(idItem, valorModal);
+        addValueArr(idItem, valorModal, true);
         valueInputModal.value = "";
         modal.style.display = "none";
         errModal.hidden = true;
-        resolve(valueInputModal.value)
+        stop = true;
       } else {
-        reject('Cancelado pelo usuário')
         errModal.hidden = false;
+        stop = true;
+      }
+
+      loop();
+
+      function loop() {
+        if (stop === true) {
+          resolve();
+          return;
+        }
+
+        setTimeout(loop, 1000);
       }
     });
   });
-};
+}
